# Conflicts:
#	src/main/resources/generatorConfig.xml
#	src/main/resources/sql-manager/distribution.sql
#	src/main/webapp/app/admMember/html/memberInfo.html

+++ src/main/java/com/distribution/dao/dateBonusHistory/mapper/DateBonusHistoryMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface DateBonusHistoryMapper {
-    int countByExample(DateBonusHistoryExample example);
+    long countByExample(DateBonusHistoryExample example);
 
     int deleteByExample(DateBonusHistoryExample example);
